backend/requirements.txt
Django==5.0
djangorestframework==3.15.2
pytest==8.3.3
flake8==7.1.1
black==24.10.0